Fix build errors related to 'ntddk.h' and 'LPCWSTR' conversion
Related to #67

Fix build errors related to 'ntddk.h' and 'LPCWSTR' conversion.

* Add the correct include path for 'ntddk.h' in `Driver/i210AVBDriver.vcxproj` for both Debug and Release configurations.
* Use the `TEXT` macro for string literals in `GUI/AVBTool.cpp` to resolve 'LPCWSTR' conversion errors.
* Use `static_cast<int>` to explicitly convert `size_t` to `int` in `AVDECC/AVDECC.cpp`, `AVTP/AVTP.cpp`, and `gPTP/gPTP.cpp`.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/zarfld/AVB-Windows/issues/67?shareId=XXXX-XXXX-XXXX-XXXX).

+++ AVDECC/AVDECC.cpp
@@ -89,7 +89,7 @@ void AVDECC::manageStreams() {
 
 void AVDECC::discoverDevices() {
     const char* discoverMessage = "DISCOVER";
-    int result = sendto(sock, discoverMessage, strlen(discoverMessage), 0, (sockaddr*)&serverAddr, sizeof(serverAddr));
+    int result = sendto(sock, discoverMessage, static_cast<int>(strlen(discoverMessage)), 0, (sockaddr*)&serverAddr, sizeof(serverAddr));
     if (result == SOCKET_ERROR) {
         std::cerr << "Send failed with error: " << WSAGetLastError() << std::endl;
         return;
@@ -97,7 +97,7 @@ void AVDECC::discoverDevices() {
 
     char buffer[1024];
     int serverAddrSize = sizeof(serverAddr);
-    result = recvfrom(sock, buffer, sizeof(buffer), 0, (sockaddr*)&serverAddr, &serverAddrSize);
+    result = recvfrom(sock, buffer, static_cast<int>(sizeof(buffer)), 0, (sockaddr*)&serverAddr, &serverAddrSize);
     if (result == SOCKET_ERROR) {
         std::cerr << "Receive failed with error: " << WSAGetLastError() << std::endl;
     } else {
