+++ gPTP/gPTP.cpp
@@ -77,7 +77,7 @@ void gPTP::synchronize() {
 
 void gPTP::sendSyncMessage() {
     const char* syncMessage = "SYNC";
-    int result = sendto(sock, syncMessage, strlen(syncMessage), 0, (sockaddr*)&serverAddr, sizeof(serverAddr));
+    int result = sendto(sock, syncMessage, static_cast<int>(strlen(syncMessage)), 0, (sockaddr*)&serverAddr, sizeof(serverAddr));
     if (result == SOCKET_ERROR) {
         std::cerr << "Send failed with error: " << WSAGetLastError() << std::endl;
     }
@@ -85,8 +85,8 @@ void gPTP::sendSyncMessage() {
 
 void gPTP::receiveSyncMessage() {
     char buffer[1024];
-    int serverAddrSize = sizeof(serverAddr);
-    int result = recvfrom(sock, buffer, sizeof(buffer), 0, (sockaddr*)&serverAddr, &serverAddrSize);
+    int serverAddrSize = static_cast<int>(sizeof(serverAddr));
+    int result = recvfrom(sock, buffer, static_cast<int>(sizeof(buffer)), 0, (sockaddr*)&serverAddr, &serverAddrSize);
     if (result == SOCKET_ERROR) {
         std::cerr << "Receive failed with error: " << WSAGetLastError() << std::endl;
     } else {
